refactor: heavily improve project-kind structure
